requirements.txt
psycopg2-binary
requests
bs4
openai
schedule